split up scatter defaults in separate files:
- require individual steps in scatter3d, scattergeo, scattergl
  defaults.
