+++ src/traces/scatter/index.js
@@ -37,6 +37,7 @@ scatter.moduleType = 'trace';
 scatter.name = 'scatter';
 scatter.categories = ['cartesian', 'symbols', 'markerColorscale', 'errorBarsOK', 'showLegend'];
 scatter.meta = {
+Scatter.supplyDefaults = require('./defaults');
     description: [
         'The scatter trace type encompasses line charts, scatter charts, text charts, and bubble charts.',
         'The data visualized as scatter point or lines is set in `x` and `y`.',
@@ -51,142 +52,6 @@ scatter.meta = {
 scatter.PTS_LINESONLY = 20;
 
 scatter.attributes = require('./attributes');
-
-var handleXYDefaults = require('./xy_defaults');
-
-scatter.supplyDefaults = function(traceIn, traceOut, defaultColor, layout) {
-    function coerce(attr, dflt) {
-        return Lib.coerce(traceIn, traceOut, scatter.attributes, attr, dflt);
-    }
-
-    var len = handleXYDefaults(traceIn, traceOut, coerce),
-        // TODO: default mode by orphan points...
-        defaultMode = len < scatter.PTS_LINESONLY ? 'lines+markers' : 'lines';
-    if(!len) {
-        traceOut.visible = false;
-        return;
-    }
-
-    coerce('text');
-    coerce('mode', defaultMode);
-
-    if(scatter.hasLines(traceOut)) {
-        scatter.lineDefaults(traceIn, traceOut, defaultColor, coerce);
-        lineShapeDefaults(traceIn, traceOut, coerce);
-        coerce('connectgaps');
-    }
-
-    if(scatter.hasMarkers(traceOut)) {
-        scatter.markerDefaults(traceIn, traceOut, defaultColor, layout, coerce);
-    }
-
-    if(scatter.hasText(traceOut)) {
-        scatter.textDefaults(traceIn, traceOut, layout, coerce);
-    }
-
-    if(scatter.hasMarkers(traceOut) || scatter.hasText(traceOut)) {
-        coerce('marker.maxdisplayed');
-    }
-
-    coerce('fill');
-    if(traceOut.fill !== 'none') {
-        scatter.fillColorDefaults(traceIn, traceOut, defaultColor, coerce);
-        if(!scatter.hasLines(traceOut)) lineShapeDefaults(traceIn, traceOut, coerce);
-    }
-
-    ErrorBars.supplyDefaults(traceIn, traceOut, defaultColor, {axis: 'y'});
-    ErrorBars.supplyDefaults(traceIn, traceOut, defaultColor, {axis: 'x', inherit: 'y'});
-};
-
-// common to 'scatter', 'scatter3d', 'scattergeo' and 'scattergl'
-scatter.lineDefaults = function(traceIn, traceOut, defaultColor, coerce) {
-    var markerColor = (traceIn.marker || {}).color;
-
-    // don't try to inherit a color array
-    coerce('line.color', (Array.isArray(markerColor) ? false : markerColor) ||
-                         defaultColor);
-    coerce('line.width');
-    coerce('line.dash');
-};
-
-function lineShapeDefaults(traceIn, traceOut, coerce) {
-    var shape = coerce('line.shape');
-    if(shape==='spline') coerce('line.smoothing');
-}
-
-// common to 'scatter', 'scatter3d', 'scattergeo' and 'scattergl'
-scatter.markerDefaults = function(traceIn, traceOut, defaultColor, layout, coerce) {
-    var isBubble = scatter.isBubble(traceIn),
-        lineColor = (traceIn.line || {}).color,
-        defaultMLC;
-
-    if(lineColor) defaultColor = lineColor;
-
-    coerce('marker.symbol');
-    coerce('marker.opacity', isBubble ? 0.7 : 1);
-    coerce('marker.size');
-
-    coerce('marker.color', defaultColor);
-    if(Colorscale.hasColorscale(traceIn, 'marker')) {
-        Colorscale.handleDefaults(
-            traceIn, traceOut, layout, coerce, {prefix: 'marker.', cLetter: 'c'}
-        );
-    }
-
-    // if there's a line with a different color than the marker, use
-    // that line color as the default marker line color
-    // mostly this is for transparent markers to behave nicely
-    if(lineColor && traceOut.marker.color!==lineColor) {
-        defaultMLC = lineColor;
-    }
-    else if(isBubble) defaultMLC = Color.background;
-    else defaultMLC = Color.defaultLine;
-
-    coerce('marker.line.color', defaultMLC);
-    if(Colorscale.hasColorscale(traceIn, 'marker.line')) {
-        Colorscale.handleDefaults(
-            traceIn, traceOut, layout, coerce, {prefix: 'marker.line.', cLetter: 'c'}
-        );
-    }
-
-    coerce('marker.line.width', isBubble ? 1 : 0);
-
-    if(isBubble) {
-        coerce('marker.sizeref');
-        coerce('marker.sizemin');
-        coerce('marker.sizemode');
-    }
-};
-
-// common to 'scatter', 'scatter3d' and 'scattergeo'
-scatter.textDefaults = function(traceIn, traceOut, layout, coerce) {
-    coerce('textposition');
-    Lib.coerceFont(coerce, 'textfont', layout.font);
-};
-
-// common to 'scatter' and 'scattergl'
-scatter.fillColorDefaults = function(traceIn, traceOut, defaultColor, coerce) {
-    var inheritColorFromMarker = false;
-
-    if(traceOut.marker) {
-        // don't try to inherit a color array
-        var markerColor = traceOut.marker.color,
-            markerLineColor = (traceOut.marker.line || {}).color;
-
-        if(markerColor && !Array.isArray(markerColor)) {
-            inheritColorFromMarker = markerColor;
-        }
-        else if(markerLineColor && !Array.isArray(markerLineColor)) {
-            inheritColorFromMarker = markerLineColor;
-        }
-    }
-
-    coerce('fillcolor', Color.addOpacity(
-        (traceOut.line || {}).color ||
-        inheritColorFromMarker ||
-        defaultColor, 0.5
-    ));
-};
 
 scatter.cleanData = function(fullData) {
     var i,
